start work on aliasing, split up ir source file
