+++ src/headers/ir.h
@@ -115,6 +115,9 @@ enum class op_type
     spill,
     spill_all,
     
+    reload_slot,
+    spill_slot,
+
     load,
 
     addrof,
@@ -266,6 +269,9 @@ enum class reg_kind
 };
 
 static constexpr u32 SIGNED_FLAG = 1 << 0;
+static constexpr u32 STORED_IN_MEM = 1 << 1;
+static constexpr u32 ALIASED = 1 << 2;
+
 
 struct Reg
 {
@@ -295,15 +301,19 @@ struct Reg
 
     b32 dirty = false;
 
-    b32 aliased = false;
-
     // NOTE: this uses absolute offsets
     // but we dont really care if they are broken by insertions during reg alloc 
     // because we only want to know when usage gap is largest
     Array<u32> usage = {};
 };
 
+struct Interloper;
+struct Function;
+struct Type;
+
 Reg make_reg(reg_kind kind,u32 size, u32 slot, b32 is_signed);
+Reg make_reg(Interloper& itl, reg_kind kind,u32 slot, const Type* type);
+b32 is_special_reg(SymSlot r);
 void destroy_reg(Reg& ir_reg);
 void print(const Reg& reg);
 
@@ -436,8 +446,6 @@ struct IrEmitter
 
 };
 
-struct Interloper;
-struct Function;
 
 void sign_extend_byte(Interloper& itl, Function& func, SymSlot dst, SymSlot src);
 void sign_extend_half(Interloper& itl, Function& func, SymSlot dst, SymSlot src);
@@ -453,9 +461,12 @@ void mov_imm(Interloper& itl, Function& func, SymSlot dst, u32 imm);
 
 void spill_rv(Interloper& itl, Function& func);
 
+void reload_slot(Interloper& itl, Function& func, const Reg& reg);
+void spill_slot(Interloper& itl, Function& func, const Reg& reg);
+
 void free_fixed_array(Interloper& itl,Function& func,SymSlot src,u32 size,u32 count);
 void free_slot(Interloper& itl,Function& func, SymSlot slot);
-struct Function;
+
 
 BlockSlot block_from_idx(u32 v);
 BlockSlot cur_block(Function& func);
